caseTransfer is moved to a separate service. Refactoring done
