+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/CaseTransferService.java
@@ -1,0 +1,131 @@
+package uk.gov.hmcts.ethos.replacement.docmosis.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.stereotype.Service;
+import uk.gov.hmcts.ecm.common.client.CcdClient;
+import uk.gov.hmcts.ecm.common.exceptions.CaseCreationException;
+import uk.gov.hmcts.ecm.common.model.ccd.CaseData;
+import uk.gov.hmcts.ecm.common.model.ccd.CaseDetails;
+import uk.gov.hmcts.ecm.common.model.ccd.items.BFActionTypeItem;
+import uk.gov.hmcts.ecm.common.model.ccd.items.DateListedTypeItem;
+import uk.gov.hmcts.ecm.common.model.ccd.items.HearingTypeItem;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.google.common.base.Strings.isNullOrEmpty;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.*;
+
+@Slf4j
+@RequiredArgsConstructor
+@Service("caseTransferService")
+public class CaseTransferService {
+
+    private final PersistentQHelperService persistentQHelperService;
+    private final CcdClient ccdClient;
+    private static final String MESSAGE = "Failed to retrieve the case for case id : ";
+
+    @Value("${ccd_gateway_base_url}")
+    private String ccdGatewayBaseUrl;
+
+    private List<CaseData> getAllCasesToBeTransferred(CaseDetails caseDetails, String userToken) {
+        try {
+            CaseData caseData = caseDetails.getCaseData();
+            List<CaseData> cases = new ArrayList<>();
+            cases.add(caseData);
+            if (caseData.getCounterClaim() != null && !caseData.getCounterClaim().trim().isEmpty()) {
+                cases.addAll(ccdClient.retrieveCasesElasticSearch(userToken,caseDetails.getCaseTypeId(), Arrays.asList(caseData.getCounterClaim())).stream().map(submitEvent -> submitEvent.getCaseData()).collect(Collectors.toList()));
+            }
+            else if (caseData.getEccCases() != null && !caseData.getEccCases().isEmpty()) {
+                List<String> counterClaimList =  caseData.getEccCases().stream().map(eccCounterClaimTypeItem -> eccCounterClaimTypeItem.getValue().getCounterClaim()).collect(Collectors.toList());
+                cases.addAll(ccdClient.retrieveCasesElasticSearch(userToken,caseDetails.getCaseTypeId(),counterClaimList).stream().map(submitEvent -> submitEvent.getCaseData()).collect(Collectors.toList()));
+            }
+            return cases;
+        }
+        catch (Exception ex) {
+            throw new CaseCreationException(MESSAGE + caseDetails.getCaseTypeId() + ex.getMessage());
+        }
+    }
+
+    public void createCaseTransfer(CaseDetails caseDetails, List<String> errors, String userToken) {
+
+        List<CaseData> caseDataList = getAllCasesToBeTransferred(caseDetails, userToken);
+        for (CaseData caseData : caseDataList) {
+
+            if (!checkBfActionsCleared(caseData)) {
+                errors.add(
+                        "There are one or more open Brought Forward actions that must be cleared before the case "
+                                + caseData.getEthosCaseReference() + " can "
+                                + "be transferred");
+            }
+
+            if (!checkHearingsNotListed(caseData)) {
+                errors.add(
+                        "There are one or more hearings that have the status Listed. These must be updated before the case "
+                                + caseData.getEthosCaseReference() + " can be transferred");
+            }
+
+            if (!errors.isEmpty()) {
+                return;
+            }
+        }
+        for (CaseData caseData : caseDataList) {
+            persistentQHelperService.sendCreationEventToSingles(
+                    userToken,
+                    caseDetails.getCaseTypeId(),
+                    caseDetails.getJurisdiction(),
+                    errors,
+                    new ArrayList<>(Collections.singletonList(caseData.getEthosCaseReference())),
+                    caseData.getOfficeCT().getValue().getCode(),
+                    caseData.getPositionTypeCT(),
+                    ccdGatewayBaseUrl,
+                    caseData.getReasonForCT(),
+                    SINGLE_CASE_TYPE,
+                    NO
+            );
+
+            caseData.setLinkedCaseCT("Transferred to " + caseData.getOfficeCT().getValue().getCode());
+            caseData.setPositionType(caseData.getPositionTypeCT());
+
+            log.info("Clearing the CT payload");
+
+            caseData.setOfficeCT(null);
+            caseData.setPositionTypeCT(null);
+            caseData.setStateAPI(null);
+        }
+    }
+
+    private boolean checkBfActionsCleared(CaseData caseData) {
+        if (caseData.getBfActions() != null) {
+            for (BFActionTypeItem bfActionTypeItem : caseData.getBfActions()) {
+                if (isNullOrEmpty(bfActionTypeItem.getValue().getCleared())) {
+                    return false;
+                }
+            }
+        }
+        return true;
+    }
+
+    private boolean checkHearingsNotListed(CaseData caseData) {
+        if (caseData.getHearingCollection() != null) {
+            for (HearingTypeItem hearingTypeItem : caseData.getHearingCollection()) {
+                if (hearingTypeItem.getValue().getHearingDateCollection() != null) {
+                    for (DateListedTypeItem dateListedTypeItem : hearingTypeItem.getValue().getHearingDateCollection()) {
+                        if (dateListedTypeItem.getValue().getHearingStatus() != null
+                                && dateListedTypeItem.getValue().getHearingStatus().equals(HEARING_STATUS_LISTED)) {
+                            return false;
+                        }
+                    }
+                }
+            }
+        }
+        return true;
+    }
+
+}
