+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/service/CaseCreationForCaseWorkerServiceTest.java
@@ -13,28 +13,17 @@ import uk.gov.hmcts.ecm.common.model.ccd.CCDRequest;
 import uk.gov.hmcts.ecm.common.model.ccd.CaseData;
 import uk.gov.hmcts.ecm.common.model.ccd.CaseDetails;
 import uk.gov.hmcts.ecm.common.model.ccd.SubmitEvent;
-import uk.gov.hmcts.ecm.common.model.ccd.items.DateListedTypeItem;
-import uk.gov.hmcts.ecm.common.model.ccd.items.HearingTypeItem;
-import uk.gov.hmcts.ecm.common.model.ccd.types.DateListedType;
-import uk.gov.hmcts.ecm.common.model.ccd.types.HearingType;
-import uk.gov.hmcts.ethos.replacement.docmosis.helpers.BFHelperTest;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.MultipleUtil;
 import uk.gov.hmcts.ethos.replacement.docmosis.utils.InternalException;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.when;
+
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.ACCEPTED_STATE;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.HEARING_STATUS_HEARD;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.HEARING_STATUS_LISTED;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.LEEDS_CASE_TYPE_ID;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.LONDON_CENTRAL_CASE_TYPE_ID;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.MULTIPLE;
 import static uk.gov.hmcts.ethos.replacement.docmosis.utils.InternalException.ERROR_MESSAGE;
 
 @RunWith(SpringJUnit4ClassRunner.class)
@@ -51,8 +40,7 @@ public class CaseCreationForCaseWorkerServiceTest {
     private SingleReferenceService singleReferenceService;
     @Mock
     private MultipleReferenceService multipleReferenceService;
-    @Mock
-    private PersistentQHelperService persistentQHelperService;
+
 
     @Before
     public void setUp() {
@@ -100,102 +88,6 @@ public class CaseCreationForCaseWorkerServiceTest {
         assertEquals(caseData.getMultipleRefNumber(),"2100005");
     }
 
-    @Test
-    public void createCaseTransfer() {
-        List<String> errors = new ArrayList<>();
-        caseCreationForCaseWorkerService.createCaseTransfer(ccdRequest.getCaseDetails(), errors, authToken);
-        assertEquals("PositionTypeCT", ccdRequest.getCaseDetails().getCaseData().getPositionType());
-        assertEquals("Transferred to " + LEEDS_CASE_TYPE_ID, ccdRequest.getCaseDetails().getCaseData().getLinkedCaseCT());
-    }
 
-    @Test
-    public void createCaseTransferECC() throws IOException {
-        List<String> errors = new ArrayList<>();
-        CaseData caseData = MultipleUtil.getCaseData("3434232323");
-        caseData.setCaseRefNumberCount("2");
-        caseData.setPositionTypeCT("PositionTypeCT1");
-        DynamicFixedListType officeCT = new DynamicFixedListType();
-        DynamicValueType valueType = new DynamicValueType();
-        valueType.setCode(LONDON_CENTRAL_CASE_TYPE_ID);
-        officeCT.setValue(valueType);
-        caseData.setOfficeCT(officeCT);
-        submitEvent.setCaseData(caseData);
-        List<SubmitEvent> submitEventList = new ArrayList<>(Collections.singletonList(submitEvent));
-        ccdRequest.getCaseDetails().getCaseData().setCounterClaim("3434232323");
-        when(ccdClient.retrieveCasesElasticSearch(anyString(),anyString(), anyList())).thenReturn(submitEventList);
-        caseCreationForCaseWorkerService.createCaseTransfer(ccdRequest.getCaseDetails(), errors, authToken);
-        assertEquals("PositionTypeCT", ccdRequest.getCaseDetails().getCaseData().getPositionType());
-        assertEquals("Transferred to " + LEEDS_CASE_TYPE_ID, ccdRequest.getCaseDetails().getCaseData().getLinkedCaseCT());
-        assertEquals("PositionTypeCT1", submitEventList.get(0).getCaseData().getPositionType());
-        assertEquals("Transferred to " + LONDON_CENTRAL_CASE_TYPE_ID, submitEventList.get(0).getCaseData().getLinkedCaseCT());
-    }
-
-    @Test
-    public void createCaseTransferMultiples() {
-        List<String> errors = new ArrayList<>();
-        ccdRequest.getCaseDetails().getCaseData().setStateAPI(MULTIPLE);
-        caseCreationForCaseWorkerService.createCaseTransfer(ccdRequest.getCaseDetails(), errors, authToken);
-        assertEquals("PositionTypeCT", ccdRequest.getCaseDetails().getCaseData().getPositionType());
-        assertEquals("Transferred to " + LEEDS_CASE_TYPE_ID, ccdRequest.getCaseDetails().getCaseData().getLinkedCaseCT());
-    }
-
-    @Test
-    public void createCaseTransferBfNotCleared() {
-        ccdRequest.getCaseDetails().getCaseData().setBfActions(BFHelperTest.generateBFActionTypeItems());
-        ccdRequest.getCaseDetails().getCaseData().getBfActions().get(0).getValue().setCleared(null);
-        List<String> errors = new ArrayList<>();
-        caseCreationForCaseWorkerService.createCaseTransfer(ccdRequest.getCaseDetails(), errors, authToken);
-        assertEquals(1, errors.size());
-        assertEquals("There are one or more open Brought Forward actions that must be cleared before the "
-                + "case " + ccdRequest.getCaseDetails().getCaseData().getEthosCaseReference() + " can be transferred", errors.get(0));
-    }
-
-    @Test
-    public void createCaseTransferHearingListed() {
-        ccdRequest.getCaseDetails().getCaseData().setHearingCollection(getHearingTypeCollection(HEARING_STATUS_LISTED));
-        List<String> errors = new ArrayList<>();
-        caseCreationForCaseWorkerService.createCaseTransfer(ccdRequest.getCaseDetails(), errors, authToken);
-        assertEquals(1, errors.size());
-        assertEquals("There are one or more hearings that have the status Listed. These must be updated "
-                + "before the case " + ccdRequest.getCaseDetails().getCaseData().getEthosCaseReference() + " can be transferred", errors.get(0));
-    }
-
-    @Test
-    public void createCaseTransferHearingListedAndBfNotCleared() {
-        ccdRequest.getCaseDetails().getCaseData().setBfActions(BFHelperTest.generateBFActionTypeItems());
-        ccdRequest.getCaseDetails().getCaseData().getBfActions().get(0).getValue().setCleared(null);
-        ccdRequest.getCaseDetails().getCaseData().setHearingCollection(getHearingTypeCollection(HEARING_STATUS_LISTED));
-        List<String> errors = new ArrayList<>();
-        caseCreationForCaseWorkerService.createCaseTransfer(ccdRequest.getCaseDetails(), errors, authToken);
-        assertEquals(2, errors.size());
-        assertEquals("There are one or more open Brought Forward actions that must be cleared before the "
-                + "case " +  ccdRequest.getCaseDetails().getCaseData().getEthosCaseReference() + " can be transferred", errors.get(0));
-        assertEquals("There are one or more hearings that have the status Listed. These must be updated "
-                + "before the case " + ccdRequest.getCaseDetails().getCaseData().getEthosCaseReference() + " can be transferred", errors.get(1));
-    }
-
-    @Test
-    public void createCaseTransferBfClearedAndNotHearingListed() {
-        ccdRequest.getCaseDetails().getCaseData().setBfActions(BFHelperTest.generateBFActionTypeItems());
-        ccdRequest.getCaseDetails().getCaseData().setHearingCollection(getHearingTypeCollection(HEARING_STATUS_HEARD));
-        List<String> errors = new ArrayList<>();
-        caseCreationForCaseWorkerService.createCaseTransfer(ccdRequest.getCaseDetails(), errors, authToken);
-        assertEquals(0, errors.size());
-    }
-
-    private List<HearingTypeItem> getHearingTypeCollection(String hearingState){
-        HearingType hearingType = new HearingType();
-        DateListedTypeItem dateListedTypeItem = new DateListedTypeItem();
-        DateListedType dateListedType = new DateListedType();
-        dateListedType.setHearingStatus(hearingState);
-        dateListedTypeItem.setId("123");
-        dateListedTypeItem.setValue(dateListedType);
-        hearingType.setHearingDateCollection(new ArrayList<>(Collections.singleton(dateListedTypeItem)));
-
-        HearingTypeItem hearingTypeItem = new HearingTypeItem();
-        hearingTypeItem.setId("1234");
-        hearingTypeItem.setValue(hearingType);
-        return new ArrayList<>(Collections.singletonList(hearingTypeItem));
-    }
 
 }
